+++ lib/src/editor/util/file_picker/file_picker_impl.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/services.dart';
 import 'package:novident_editor/src/editor/util/file_picker/file_picker_service.dart';
 import 'package:file_picker/file_picker.dart' as fp;
 
@@ -14,9 +15,6 @@ class FilePicker implements FilePickerService {
     fp.FileType type = fp.FileType.any,
     List<String>? allowedExtensions,
     Function(fp.FilePickerStatus p1)? onFileLoading,
-    bool allowMultiple = false,
-    bool withData = false,
-    bool withReadStream = false,
     bool lockParentWindow = false,
   }) async {
     final result = await fp.FilePicker.pickFiles(
@@ -25,9 +23,6 @@ class FilePicker implements FilePickerService {
       type: type,
       allowedExtensions: allowedExtensions,
       onFileLoading: onFileLoading,
-      allowMultiple: allowMultiple,
-      withData: withData,
-      withReadStream: withReadStream,
       lockParentWindow: lockParentWindow,
     );
     return FilePickerResult(result?.files ?? []);
@@ -35,14 +30,18 @@ class FilePicker implements FilePickerService {
 
   @override
   Future<String?> saveFile({
+    required Uint8List bytes,
+    required String fileName,
     String? dialogTitle,
-    String? fileName,
     String? initialDirectory,
     fp.FileType type = fp.FileType.any,
     List<String>? allowedExtensions,
     bool lockParentWindow = false,
+    dynamic Function(fp.FilePickerStatus)? onFileLoading,
   }) {
     return fp.FilePicker.saveFile(
+      bytes: bytes,
+      onFileLoading: onFileLoading,
       dialogTitle: dialogTitle,
       fileName: fileName,
       initialDirectory: initialDirectory,