+++ lib/src/editor/util/file_picker/file_picker_service.dart
@@ -1,10 +1,11 @@
-import 'package:file_picker/file_picker.dart';
+import 'package:file_picker/file_picker.dart' as fp;
+import 'package:flutter/foundation.dart';
 
 class FilePickerResult {
   const FilePickerResult(this.files);
 
   /// Picked files.
-  final List<PlatformFile> files;
+  final List<fp.PlatformFile> files;
 }
 
 /// Abstract file picker as a service to implement dependency injection.
@@ -17,23 +18,22 @@ abstract class FilePickerService {
   Future<FilePickerResult?> pickFiles({
     String? dialogTitle,
     String? initialDirectory,
-    FileType type = FileType.any,
+    fp.FileType type = fp.FileType.any,
     List<String>? allowedExtensions,
-    Function(FilePickerStatus)? onFileLoading,
-    bool allowMultiple = false,
-    bool withData = false,
-    bool withReadStream = false,
+    Function(fp.FilePickerStatus p1)? onFileLoading,
     bool lockParentWindow = false,
   }) async =>
       throw UnimplementedError('pickFiles() has not been implemented.');
 
   Future<String?> saveFile({
+    required Uint8List bytes,
+    required String fileName,
     String? dialogTitle,
-    String? fileName,
     String? initialDirectory,
-    FileType type = FileType.any,
+    fp.FileType type = fp.FileType.any,
     List<String>? allowedExtensions,
     bool lockParentWindow = false,
+    dynamic Function(fp.FilePickerStatus)? onFileLoading,
   }) async =>
       throw UnimplementedError('saveFile() has not been implemented.');
 }